requirements.txt
watchdog==3.0.0
requests==2.31.0
pyinstaller==6.3.0
discord.py==1.7.3
